Factor out the common traits of the various GroupStrategies and turn them into "sub strategies". Add a common base class for the public strategies, that plug the various pieces together and allows flexible ordering of the sub-pieces.
Add a "SmartJdbiGroupStrategy" that can handle both SqlObject queries
and regular queries (if the statement locator sets the correct values
on the context).

The statement locator can now set a third value (_jmx_class), which
then is interpreted by the SmartJdbiGroupStrategy similar to the
SqlJdbiGroupStrategy (and gives an unified Jmx tree for all the DAOs).